added exceptional handler
