+++ src/main/java/com/jpademo/JpaDemo/exception/BookExceptionHandler.java
@@ -30,28 +30,4 @@ public class BookExceptionHandler {
         return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
     }
 
-    @Component
-    public static class AuthEntryPointJwt implements AuthenticationEntryPoint {
-
-      private static final Logger logger = LoggerFactory.getLogger(AuthEntryPointJwt.class);
-
-      @Override
-      public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException)
-          throws IOException, ServletException {
-        logger.error("Unauthorized error: {}", authException.getMessage());
-
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-
-        final Map<String, Object> body = new HashMap<>();
-        body.put("status", HttpServletResponse.SC_UNAUTHORIZED);
-        body.put("error", "Unauthorized");
-        body.put("message", authException.getMessage());
-        body.put("path", request.getServletPath());
-
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.writeValue(response.getOutputStream(), body);
-      }
-
-    }
 }
